refactor: move all SDAM events into a common file
