+++ lib/core/sdam/topology.js
@@ -4,7 +4,7 @@ const ServerDescription = require('./server_description').ServerDescription;
 const ServerType = require('./common').ServerType;
 const TopologyDescription = require('./topology_description').TopologyDescription;
 const TopologyType = require('./common').TopologyType;
-const monitoring = require('./monitoring');
+const events = require('./events');
 const Server = require('./server').Server;
 const relayEvents = require('../utils').relayEvents;
 const ReadPreference = require('../topologies/read_preference');
@@ -272,12 +272,12 @@ class Topology extends EventEmitter {
     stateTransition(this, STATE_CONNECTING);
 
     // emit SDAM monitoring events
-    this.emit('topologyOpening', new monitoring.TopologyOpeningEvent(this.s.id));
+    this.emit('topologyOpening', new events.TopologyOpeningEvent(this.s.id));
 
     // emit an event for the topology change
     this.emit(
       'topologyDescriptionChanged',
-      new monitoring.TopologyDescriptionChangedEvent(
+      new events.TopologyDescriptionChangedEvent(
         this.s.id,
         new TopologyDescription(TopologyType.Unknown), // initial is always Unknown
         this.s.description
@@ -377,7 +377,7 @@ class Topology extends EventEmitter {
           this.s.servers.clear();
 
           // emit an event for close
-          this.emit('topologyClosed', new monitoring.TopologyClosedEvent(this.s.id));
+          this.emit('topologyClosed', new events.TopologyClosedEvent(this.s.id));
 
           stateTransition(this, STATE_CLOSED);
           this.emit('close');
@@ -548,7 +548,7 @@ class Topology extends EventEmitter {
     // emit monitoring events for this change
     this.emit(
       'serverDescriptionChanged',
-      new monitoring.ServerDescriptionChangedEvent(
+      new events.ServerDescriptionChangedEvent(
         this.s.id,
         serverDescription.address,
         previousServerDescription,
@@ -561,7 +561,7 @@ class Topology extends EventEmitter {
 
     this.emit(
       'topologyDescriptionChanged',
-      new monitoring.TopologyDescriptionChangedEvent(
+      new events.TopologyDescriptionChangedEvent(
         this.s.id,
         previousTopologyDescription,
         this.s.description
@@ -786,7 +786,7 @@ function destroyServer(server, topology, options, callback) {
   server.destroy(options, () => {
     topology.emit(
       'serverClosed',
-      new monitoring.ServerClosedEvent(topology.s.id, server.description.address)
+      new events.ServerClosedEvent(topology.s.id, server.description.address)
     );
 
     if (typeof callback === 'function') {
@@ -821,7 +821,7 @@ function randomSelection(array) {
 function createAndConnectServer(topology, serverDescription, connectDelay) {
   topology.emit(
     'serverOpening',
-    new monitoring.ServerOpeningEvent(topology.s.id, serverDescription.address)
+    new events.ServerOpeningEvent(topology.s.id, serverDescription.address)
   );
 
   const server = new Server(serverDescription, topology.s.options, topology);
@@ -1061,7 +1061,7 @@ function srvPollingHandler(topology) {
 
     topology.emit(
       'topologyDescriptionChanged',
-      new monitoring.TopologyDescriptionChangedEvent(
+      new events.TopologyDescriptionChangedEvent(
         topology.s.id,
         previousTopologyDescription,
         topology.s.description
